# server/http/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Report API",
    "version": "1.0.0",
    "description": "A REST API for generating reports."
  },
...
  "paths": {
...
    "/tutorials": {
      "post": {
        "operationId": "createTutorial",
        "tags": [
          "tutorials"
        ],
        "description": "Create an new Tutorial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "OTHERS",
                      "MOVEMENT",
                      "REGISTER",
                      "FINANCIAL"
                    ]
                  }
                },
                "required": [
                  "title",
                  "category"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getTutorials",
        "tags": [
          "tutorials"
        ],
        "description": "List all Tutorials",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      },
                      "category": {
                        "type": "string",
                        "enum": [
                          "OTHERS",
                          "MOVEMENT",
                          "REGISTER",
                          "FINANCIAL"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "category"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tutorials/{id}": {
      "get": {
        "operationId": "getTutorial",
        "tags": [
          "tutorials"
        ],
        "description": "Get a tutorial by ID",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string",
                      "nullable": true
                    },
                    "category": {
                      "type": "string",
                      "enum": [
                        "OTHERS",
                        "MOVEMENT",
                        "REGISTER",
                        "FINANCIAL"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "content",
                    "category"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateTutorial",
        "tags": [
          "tutorials"
        ],
        "description": "Update a tutorial by ID",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "OTHERS",
                      "MOVEMENT",
                      "REGISTER",
                      "FINANCIAL"
                    ]
                  },
                  "content": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTutorial",
        "tags": [
          "tutorials"
        ],
        "description": "Delete a tutorial by ID",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}